clean up data modules
